Renamed src/user/dto to dtos, created subdirectories for user and motorcycle, and initialized CreateMotorcycleDto.
