+++ src/user/dtos/motorcycle/create-motorcycle.dto.ts
@@ -1,0 +1,10 @@
+export class CreateMotorcycleDto {
+  licensePlate: string;
+  brande: string;
+  year: string;
+  model: string;
+  color: string;
+  isActive: boolean;
+  owner: string;
+  driver: string;
+}
