+++ src/user/dtos/user/response-user.dto.ts
@@ -1,5 +1,5 @@
 import { Role } from 'src/common/role/roles.enum';
-import { DeliveryMan, User } from '../entities/user.entity';
+import { DeliveryMan, User } from '../../entities/user.entity';
 import { ResponseVoucherDto } from 'src/voucher/dto/response-voucher.dto';
 import { Tip } from 'src/tip/entities/tip.entity';
 import { SmallResponseWorkTime } from 'src/work-time/types/small-response-work-time.type';
