+++ src/user/user.service.ts
@@ -9,10 +9,10 @@ import {
 import { Repository } from 'typeorm';
 import { DeliveryMan, User } from './entities/user.entity';
 import { InjectRepository } from '@nestjs/typeorm';
-import { CreateUserDto } from './dto/create-user.dto';
+import { CreateUserDto } from './dtos/user/create-user.dto';
 import { HashingService } from 'src/common/hashing/hashing.service';
-import { UpdateUserDto } from './dto/update-user.dto';
-import { UpdatePasswordDto } from './dto/update-password.dto';
+import { UpdateUserDto } from './dtos/user/update-user.dto';
+import { UpdatePasswordDto } from './dtos/user/update-password.dto';
 import { RoleService } from 'src/common/role/role.service';
 import { Role, Role as RoleEnum, roles } from 'src/common/role/roles.enum';
 import { essencial, full } from './data/relations/user';
